Fixed Everything Except the Size Error
Small errors here and there, but nothing major. I tried every method with a full list and with an empty list. For the size problem, look at the CellList copy constructor;  I tried a debugging by adding Sout(this.size) in most CellLists methods, the problem seems to be in copy constructor OR equals.